Refactoring middleware location, adding script to create first users (admins)
