+++ ElevateEatsAPI/middleware/apiKeyAuth.js
